Some fixes, read Desc for more
Disabled fuel system for now because it throws too many errors, but I'd like to activate it again as soon as I figured it out. Implemented little fix for Profile dialog so there wont be any "dialog id already exists" errors anymore. Implemented new functionality for bus job, now only one person can drive a lane to avoid bugs with the checkpoints and so on.